Linked all the menu items.

+++ SQL/products.create.sql
@@ -2,7 +2,7 @@ CREATE TABLE products (
     product_id SERIAL PRIMARY KEY,
     product_name varchar(255) NOT NULL,
     product_description varchar(255) NOT NULL,
-    k_percent decimal(10,2) NOT NULL,
-    p_percent decimal(10,2) NOT NULL,
-    n_percent decimal(10,2) NOT NULL
+    k_percent int NOT NULL,
+    p_percent int NOT NULL,
+    n_percent int NOT NULL
 );
